redis-caching/requirements.txt
Flask
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
Flask-Caching
python-dotenv
requests
gunicorn
redis
